refactoring
moved code to new class SpringEngine
